+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [#80](https://github.com/itk-dev/devops_itksites/pull/80) 5566: Service agreements
+  - Add security contract entity with crud controller
+  - Add Abstract full crud controller and extend on it in some cases
+  - Add economics service and sync action/command for service agreement synchronization
 - [#81](https://github.com/itk-dev/devops_itksites/pull/81) 5564: Asset Mapper migration
   - Add Symfony Asset Mapper bundle and importmap
 - Add Renovate auto-patch + auto-release pipeline (Phase 1 fork validation)